+++ cmake/Debug.cmake
@@ -1,5 +1,3 @@
 set(CMAKE_CXX_FLAGS ${CMAKE_CXX_FLAGS_DEBUG})
 
-# -DDEBUG: #define DEBUG
-add_compile_definitions(DEBUG)
 add_compile_definitions(_DEBUG)