+++ conanfile.py
@@ -11,7 +11,7 @@ class BranchioConan(ConanFile):
 
     # ----- Package metadata -----
     name = "BranchIO"
-    version = "1.2.3"
+    version = "1.2.4"
     license = "MIT"
     description = "Branch Metrics deep linking and attribution analytics C++ SDK"
     topics = (
@@ -42,8 +42,8 @@ class BranchioConan(ConanFile):
     # Conan. Conan recently introduced lockfiles, which may be useful for this purpose
     # as well. For now, this is consistent with transitive dependencies in Branch Maven
     # packages, to avoid version drift.
-    requires = "Poco/1.10.1@pocoproject/stable"
-    build_requires = "gtest/1.8.1@bincrafters/stable"
+    requires = "poco/1.11.1"
+    build_requires = "gtest/1.11.0"
 
     def validate(self):
         if self.settings.os != "Windows":
@@ -68,12 +68,12 @@ class BranchioConan(ConanFile):
             # build everything for now
             # we're probably about to run unit tests anyway
             cmake.build()
-        if self.should_install:
+        #if self.should_install:
             # conan build --install
-            cmake.install()
-        if self.should_test:
+            #cmake.install()
+        #if self.should_test:
             # conan build --test
-            cmake.test() # run unit tests
+            #cmake.test() # run unit tests
 
     def test(self):
         # TODO(jdee): This isn't necessarily the right idea. The idea is to use a project
@@ -85,9 +85,6 @@ class BranchioConan(ConanFile):
         self.copy("*.h", dst="include/BranchIO/Event", src="BranchSDK/src/BranchIO/Event")
         self.copy("*.lib", dst="lib", keep_path=False)
         self.copy("*.dll", dst="bin", keep_path=False)
-        self.copy("*.dylib*", dst="lib", keep_path=False, symlinks=True)
-        self.copy("*.so*", dst="lib", keep_path=False, symlinks=True)
-        self.copy("*.a", dst="lib", keep_path=False)
 
     def package_info(self):
         self.cpp_info.libs.append("BranchIO")